split 3d and 2d gait analysis
